+++ .sbtopts
@@ -1,0 +1,1 @@
+-Dsbt.io.implicit.relative.glob.conversion=allow
