+++ modules/integrations/cats/src/main/scala/tethys/cats/instances.scala
