+++ modules/integrations/enumeratum/src/main/scala/tethys/enumeratum/TethysEnum.scala
