+++ modules/integrations/enumeratum/src/main/scala/tethys/enumeratum/TethysKeyEnum.scala
