+++ modules/integrations/enumeratum/src/main/scala/tethys/enumeratum/TethysValueEnum.scala
@@ -1,0 +1,41 @@
+package tethys.enumeratum
+
+import _root_.enumeratum.values.*
+import tethys.readers.KeyReader
+import tethys.writers.KeyWriter
+import tethys.{JsonReader, JsonWriter}
+
+sealed trait TethysValueEnum[
+    ValueType,
+    EntryType <: ValueEnumEntry[ValueType]
+] { self: ValueEnum[ValueType, EntryType] =>
+  implicit def tethysReader: JsonReader[EntryType]
+  implicit def tethysWriter: JsonWriter[EntryType]
+}
+
+trait StringTethysEnum[E <: StringEnumEntry]
+    extends TethysValueEnum[String, E] { self: ValueEnum[String, E] =>
+  implicit val tethysReader: JsonReader[E] = Enumeratum.valueReader(this)
+  implicit val tethysWriter: JsonWriter[E] = Enumeratum.valueWriter(this)
+  implicit val tethysKeyReader: KeyReader[E] =
+    Enumeratum.keyReader(this)(_.withValueOpt)
+  implicit val tethysKeyWriter: KeyWriter[E] = Enumeratum.keyWriter(_.value)
+}
+
+trait IntTethysEnum[E <: IntEnumEntry] extends TethysValueEnum[Int, E] {
+  self: ValueEnum[Int, E] =>
+  implicit val tethysReader: JsonReader[E] = Enumeratum.valueReader(this)
+  implicit val tethysWriter: JsonWriter[E] = Enumeratum.valueWriter(this)
+}
+
+trait LongTethysEnum[E <: LongEnumEntry] extends TethysValueEnum[Long, E] {
+  self: ValueEnum[Long, E] =>
+  implicit val tethysReader: JsonReader[E] = Enumeratum.valueReader(this)
+  implicit val tethysWriter: JsonWriter[E] = Enumeratum.valueWriter(this)
+}
+
+trait ShortTethysEnum[E <: ShortEnumEntry] extends TethysValueEnum[Short, E] {
+  self: ValueEnum[Short, E] =>
+  implicit val tethysReader: JsonReader[E] = Enumeratum.valueReader(this)
+  implicit val tethysWriter: JsonWriter[E] = Enumeratum.valueWriter(this)
+}
