+++ modules/integrations/enumeratum/src/test/scala/tethys/enumeratum/Data.scala
