+++ modules/integrations/enumeratum/src/test/scala/tethys/enumeratum/Direction.scala
